+++ features/review/index.ts
@@ -12,3 +12,6 @@ export {
 export { ReviewListCard } from './components/review-list-card';
 export { SoftSkillsCard } from './components/softskills-card';
 export { TagCard } from './components/tag-card';
+export type { CreateReviewRequest, Review, ReviewFormData } from './types/review';
+export type { Tag, ReceivedTags, ReceivedTagCount } from './types/tag';
+export type { Spectrum, SpectrumInfo, SpectrumStats, SpectrumWithValue } from './types/spectrum';
